After adding photos fails, don't try to unlink the old file unless it exists
git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery@437 57fcd75e-5312-0410-8df3-f5eb6fbb1595

+++ classes/Album.php
@@ -166,7 +166,9 @@ class Album {
 		$item = new AlbumItem();
 		$err = $item->setPhoto($dir, $name, $tag, $this->fields["thumb_size"]);
 		if ($err) {
-			unlink("$dir/$name.$tag");
+			if (file_exists("$dir/$name.$tag")) {
+				unlink("$dir/$name.$tag");
+			}
 			return $err;
 		}
 		$this->photos[] = $item;
